imporoving location render logic
